move already_analyzed back to surface.py, we need a more elegant way to organize this. Also get_symmetrically_distinct/equivalent_indices can now return hkil
